Refactor brands pagination

+++ MilaIrisWebParent/MilaIrisBackEnd/src/main/java/kz/milairis/admin/category/CategoryController.java
@@ -1,7 +1,6 @@
 package kz.milairis.admin.category;
 
 import kz.milairis.admin.FileUploadUtil;
-import kz.milairis.admin.user.UserService;
 import kz.milairis.common.entity.Category;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.query.Param;
@@ -32,7 +31,8 @@ public class CategoryController {
 	@GetMapping("/categories/page/{pageNum}")
 	public String listByPage(@PathVariable(name = "pageNum") int pageNum,
 							 @Param("sortDir") String sortDir,
-							 @Param("keyword") String keyword, Model model) {
+							 @Param("keyword") String keyword,
+							 Model model) {
 		if (sortDir ==  null || sortDir.isEmpty()) {
 			sortDir = "asc";
 		}
@@ -40,7 +40,7 @@ public class CategoryController {
 		CategoryPageInfo pageInfo = new CategoryPageInfo();
 		List<Category> listCategories = service.listByPage(pageInfo, pageNum, sortDir, keyword);
 
-		long startCount = (pageNum - 1) * CategoryService.ROOT_CATEGORIES_PER_PAGE + 1;
+		long startCount = (long) (pageNum - 1) * CategoryService.ROOT_CATEGORIES_PER_PAGE + 1;
 		long endCount = startCount + CategoryService.ROOT_CATEGORIES_PER_PAGE - 1;
 		if (endCount > pageInfo.getTotalElements()) {
 			endCount = pageInfo.getTotalElements();
@@ -149,3 +149,4 @@ public class CategoryController {
 		exporter.export(listCategories, response);
 	}
 }
+
